# API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: SIAR Platform ERP/OMS/WMS API
  description: Enterprise-grade Order Management, Inventory Control, and Fulfillment API
  version: 1.0.0
  contact:
    name: SIAR Digital Support
    email: support@siardigital.com
  license:
    name: Proprietary
    url: https://www.siardigital.com
servers:
  - url: http://localhost:5000/api
    description: Development Server
  - url: https://erp.siardigital.com/api
    description: Production Server
...
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create new order
      description: Create an order from any marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: integer
                marketplace:
                  type: string
                  enum: [Shopify, Amazon, TikTok, Walmart, eBay, Etsy]
                marketplace_order_id:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                total_amount:
                  type: number
      responses:
        '201':
          description: Order created
        '400':
          description: Invalid order data

    get:
      tags:
        - Orders
      summary: List orders
      description: Get orders with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: marketplace
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Orders list
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Retrieve complete order information
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
        '404':
          description: Order not found
  /orders/{orderId}/status:
    post:
      tags:
        - Orders
      summary: Update order status
      description: Change order status with audit trail
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_status:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Status updated

# =====================================================================
# INVENTORY MANAGEMENT ENDPOINTS (Core - 12+)
# =====================================================================
  /orders/warehouses:
    get:
      tags:
        - Inventory
      summary: List warehouses
      responses:
        '200':
          description: Warehouses list

    post:
      tags:
        - Inventory
      summary: Create warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                warehouse_name:
                  type: string
                location:
                  type: string
                capacity:
                  type: integer
      responses:
        '201':
          description: Warehouse created
  /orders/skus:
    get:
      tags:
        - Inventory
      summary: List SKUs
      responses:
        '200':
          description: SKUs list

    post:
      tags:
        - Inventory
      summary: Create SKU
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sku_name:
                  type: string
                sku_code:
                  type: string
                description:
                  type: string
                category:
                  type: string
      responses:
        '201':
          description: SKU created
  /orders/inventory/{skuId}:
    get:
      tags:
        - Inventory
      summary: Get SKU inventory
      parameters:
        - name: skuId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Inventory details
...
security:
  - BearerToken: []

# =====================================================================
# COMPONENTS & SCHEMAS
# =====================================================================
...